refactor(backend): isolate direct selfhood prompts
- Move direct selfhood/origin prompt contracts into direct_prompt_selfhood.py without changing prompt text

- Keep direct_prompts.py focused on system-message assembly and import the new contract helpers

- Update the runtime cleanup audit with the #533 boundary

Refs #533
